+++ provision/centos.sh
@@ -19,9 +19,6 @@ export JAIL_START_EXTRA="allow.mount
 export JAIL_CONF_EXTRA='
 		allow.raw_sockets;'
 export JAIL_FSTAB="
-$ZFS_JAIL_MNT/centos/dev $ZFS_JAIL_MNT/centos/compat/linux/dev nullfs rw  0 0
-tmpfs     $ZFS_JAIL_MNT/centos/compat/linux/dev/shm tmpfs     rw,size=1g,mode=1777  0 0
-fdescfs   $ZFS_JAIL_MNT/centos/compat/linux/dev/fd  fdescfs   rw,linrdlnk 0 0
 linprocfs $ZFS_JAIL_MNT/centos/compat/linux/proc    linprocfs rw  0 0
 linsysfs  $ZFS_JAIL_MNT/centos/compat/linux/sys     linsysfs  rw  0 0
 #/tmp      $ZFS_JAIL_MNT/centos/compat/linux/tmp     nullfs    rw  0 0
@@ -39,6 +36,7 @@ for _fs in dev proc sys tmp home; do
 done
 chmod 777 "$STAGE_MNT/compat/linux/tmp"
 assure_devfs_linux_ruleset
+configure_linux_devfs
 start_staged_jail centos
 install_centos
 promote_staged_jail centos